Revert "Merge branch 'itamarst/correct-protocol-handshakes' into itamarst/environment-end-to-end-tests"
This reverts commit e02b9bfe7ca827389402be3753c2b55754aba40c, reversing
changes made to 1a981804db0228538d2e10d2bda3ae00d538d9fc.
