+++ quark/mdk_runtime_fake.q
@@ -99,14 +99,6 @@ namespace mdk_runtime {
         }
 
         @doc("""
-        Return number of sent messages we haven't looked at yet with
-        expectTextMessage().
-        """)
-        int uninspectedSentMessages() {
-            return sent.size() - expectIdx;
-        }
-
-        @doc("""
         Check that a message has been sent via this actor.
         """)
         String expectTextMessage() {
@@ -120,7 +112,7 @@ namespace mdk_runtime {
                 expectIdx = expectIdx + 1;
                 return msg;
             }
-            panic("no remaining message found");
+            Context.runtime().fail("no remaining message found");
             return "unreachable";
         }
     }
