+++ quark/protocol-1.0.q
@@ -397,7 +397,6 @@ namespace mdk_protocol {
         WSClient _wsclient;
         String _node_id;
         OperationalEnvironment _environment;
-        Actor _websocket;
 
         OpenCloseSubscriber(WSClient client, String node_id, OperationalEnvironment environment) {
             self._wsclient = client;
@@ -432,7 +431,11 @@ namespace mdk_protocol {
         }
 
         void onWSConnected(Actor websocket) {
-            _websocket = websocket;
+            // Send Open message to the server:
+            Open open = new Open();
+            open.nodeId = self._node_id;
+            open.environment = _environment;
+            self._dispatcher.tell(self, open.encode(), websocket);
         }
 
         void onPump() {}
@@ -440,12 +443,6 @@ namespace mdk_protocol {
         // WebSocket message handlers:
         void onOpen() {
             // Should assert version here ...
-
-            // Send Open message to the server:
-            Open open = new Open();
-            open.nodeId = self._node_id;
-            open.environment = _environment;
-            self._dispatcher.tell(self, open.encode(), _websocket);
         }
 
         void onClose(Close close) {
